signal: call per file
c_src/sys/alcove_signal.c has a function to map signal numbers to
constants used by the signal handler in the event loop.
